backend/requirements-backend.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
drf-spectacular==0.28.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
boto3==1.42.55
clickhouse-driver==0.2.9
kafka-python==2.1.2
pymongo==4.9.2
python-dotenv==1.0.1
requests==2.32.3
